+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 4.3.3
 
 * More robust error handling when checking/execting scripts
+* Bug fix for when setting other pieces global/indexed/mutal scripts
 
 ## 4.3.2